Create UseCaseOne first endpoint
This endpoint simulates the start of a new charging session.It realizes a get request. It accepts as a parameter
a chargerId (the chargerId of the vehicle that requests the charging session) and it returns a random charging point
with the same connection protocol along with all the station, the energy provider and the charger data.